Fix incorrect base type for price filler

+++ src/GenFu/Fillers/NumericFillers.cs
@@ -102,7 +102,7 @@ namespace GenFu
         }
     }
 
-    public class PriceFiller : PropertyFiller<int>
+    public class PriceFiller : PropertyFiller<decimal>
     {
         private const int _maxPrice = 1000;
